refactor: move exercises list to component
